extract authOptions
